lib/lingo/database.rb: Provide alternative storage backends.
SDBM is fast and comes with Ruby's stdlib, but has a limit for the record size. GDBM doesn't have that limit, but is slower. Depot (from QDBM) would be faster than GDBM, but is complicated to install. SQLite would be the most flexible, but is way too slow. CDB, finally, is the fastest from all, but isn't available on Windows (though that can possibly be changed).

So, for now, we prefer CDB (if available) over GDBM (if available) and fall back to SDBM. Use of CDB may be prevented by setting the LINGO_NO_CDB environment variable; for GDBM it's LINGO_NO_GDBM accordingly.